added upov endpoint

+++ backend/camellia/src/main/java/com/camellia/services/characteristics/CharacteristicService.java
@@ -24,6 +24,6 @@ public class CharacteristicService {
     }
 
     public Characteristic getCharacteristicById(long id) {
-        return repository.findById((long) id);
+        return repository.findById(id);
     }
 }
